[fixed] harvester radials
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk/MMOCoreORB@2397 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/packets/object/ObjectMenuResponse.h
@@ -113,7 +113,7 @@ public:
 			return root->getItem(index);
 	}
 
-	RadialMenuItem* getRadialItemByRadialID(int id) {
+	RadialMenuItem* getRadialItemByRadialID(uint8 id) {
 		if (id == root->getRadialID())
 			return root;
 		else
